feat: expanded networks, refactored directory structure, and removed defaults
