--- typescript/agentkit/src/action-providers/onramp/utils/networkConversion.ts
@@ -1,10 +1,0 @@
-export const convertNetworkIdToOnrampNetworkId = (networkId: string): string | null => {
-  switch (networkId) {
-    case "base-mainnet":
-      return "base";
-    case "base-sepolia":
-      return "base-sepolia";
-    default:
-      return null;
-  }
-};
